[#77]: Move DescriptorSets to Shader struct
